requirements.txt
# requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-SQLAlchemy==3.1.1
Flask-RESTful==0.3.10
Flask-Talisman==1.1.0
Flask-Limiter==4.0.1
PyOTP==2.9.0
bcrypt==4.2.0
python-dotenv==1.0.1
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
gunicorn==23.0.0
redis==5.0.8
cryptography==43.0.1
alembic==1.13.2